Revert "feat: Wishlist"
