+++ src/Context.js
@@ -1,34 +1,28 @@
 import React from 'react'
-import { useReducer, useEffect } from 'react';
+import { useReducer } from 'react';
 import { createContext } from 'react';
 import stateReducer from './Reducer';
 
+export const StateContext=createContext();
 
-export const StateContext = createContext();
-
-const Context = ({ children }) => {
-
-  const [state, dispatch] = useReducer(stateReducer, {
-    products: [],
-    categories: [],
-    sortBy: null,
-    priceRange: 0,
-    rating: 0,
-    sortByFilters: {
-      sortByCategory: [],
-      sortByBrands: []
-    },
-    subCategory: null,
-    cart: [],
-    wishlist: []
-  })
-
-
-
+const Context = ({children}) => {
+    const [state,dispatch]=useReducer(stateReducer,{
+        products:[],
+        categories:[],
+        sortBy:null,
+        priceRange:0,
+        rating:0,
+        sortByFilters:{
+            sortByCategory:[],
+            sortByBrands:[]
+        },
+        subCategory:null,
+        cart:[]
+    })
 
   return (
-    <StateContext.Provider value={{ state, dispatch }}>
-      {children}
+    <StateContext.Provider value={{state,dispatch}}>
+        {children}
     </StateContext.Provider>
   )
 }
