+++ src/Reducer.js
@@ -30,7 +30,6 @@ const stateReducer = (state, action) => {
             subCategory: null
         }
         case 'SET_CART':return {...state,cart:action.payload}
-        case 'SET_WISHLIST':return {...state,wishlist:action.payload}
         default: return state
     }
