+++ src/components/Header/Header.css
@@ -1,10 +1,6 @@
 .navigation-container{
     background-color: var(--white);
     border-bottom: 0.5px solid var(--grey);
-    position: sticky;
-    top:39px;
-    left: 0px;
-    z-index: 2;
 }
 .lib-name-1,.lib-name-2{
     font-family: 'Cormorant Upright', serif;
