+++ src/components/ProductCard/ProductCard.js
@@ -2,18 +2,16 @@ import React from 'react'
 import { useContext } from 'react'
 import { useState } from 'react'
 import { useNavigate } from 'react-router-dom';
-import { addToCart, updateProductQty } from '../../CartServices'
+import { addToCart } from '../../CartServices'
 import { StateContext } from '../../Context'
-import { addToWishlist, handleAddToWishlist, removeFromWishlist } from '../../WishlistServices';
 
 import './ProductCard.css'
 
 const ProductCard = ({ item }) => {
   const encodedToken = localStorage.getItem("token")
-  const { state,dispatch } = useContext(StateContext)
+  const { state } = useContext(StateContext)
   let navigate = useNavigate();
   const [cartButtonText, setCartButtonText] = useState('ADD TO CART')
-  const [heartColor,setHeartColor]=useState('#F3C5C5')
   const handleAddTOCart = () => {
     const isItemPresent = state.cart.find(itemInCart => itemInCart._id === item._id)
     
@@ -22,10 +20,6 @@ const ProductCard = ({ item }) => {
         addToCart(item, encodedToken)
         setCartButtonText('GO TO CART')
       } else {
-        const isItemPresentInWishList = state.wishlist.find(itemInWishlist => itemInWishlist._id === item._id)
-        if(isItemPresentInWishList!==undefined){
-          updateProductQty(item._id, encodedToken, dispatch, "increment")
-        }
         setCartButtonText("GO TO CART")
       }
     } else {
@@ -33,8 +27,6 @@ const ProductCard = ({ item }) => {
     }
   }
 
-  
-  
 
   return (
     <div class="card__container flex-vt">
@@ -50,9 +42,7 @@ const ProductCard = ({ item }) => {
         <span><i class="fa fa-star"></i></span>
       </div>
       <div class="card__footer flex-hz">
-
-      {state.wishlist.includes(item) ? <>  <div><button class="btn btn-primary card__btn-primary" onClick={()=>removeFromWishlist(item._id,encodedToken,dispatch)}>Remove</button></div></>:<>  <div><i style={{color: `${heartColor}`}} class="fa fa-heart card__icon" onClick={()=> handleAddToWishlist(state.wishlist,item, encodedToken)}></i></div></>}
-      
+        <div><i class="fa fa-heart card__icon"></i></div>
         <button class="btn btn-primary card__btn-primary" onClick={handleAddTOCart}>{cartButtonText}</button>
       </div>
     </div>
