+++ src/components/Sidebar/Sidebar.css
@@ -5,7 +5,7 @@
   grid-area: sidebar;
   width: 30rem;
   box-shadow: rgba(100, 100, 111, 0.2) 0px 7px 29px 0px;
-  z-index: -1;
+  z-index: 2;
 }
 
 .filter-sidebar-items {
