--- src/pages/WishlistPage/WishlistPage.css
@@ -1,4 +1,0 @@
-.card-icon-wishlist {
-    color: var(--primary-text-color);
-  }
-
